+++ app/src/field/RenderField.tsx
@@ -8,18 +8,17 @@ interface FieldProps extends WrappedFieldProps {
     disabled: boolean
 }
 
-export class RenderField extends React.Component<FieldProps> {
-    render() {
-        const { className, label, type, input, disabled, meta: {touched, error} } = this.props
-        return (
-            <React.Fragment>
-                { type === 'textarea' ? 
-                <textarea className={className} placeholder={label} {...input}/>
-                :
-                <input className={className} placeholder={label} {...input} type={type}/>
-                }
-                { touched && error && <span>{error}</span> }
-            </React.Fragment>
-        )
-    }
+export const RenderField: React.FC<FieldProps> = (props: FieldProps) => {
+    const { className, label, type, input, disabled, meta: {touched, error} } = props
+
+    return (
+        <React.Fragment>
+            { type === 'textarea' ? 
+            <textarea className={className} placeholder={label} {...input}/>
+            :
+            <input className={className} placeholder={label} {...input} type={type}/>
+            }
+            { touched && error && <span>{error}</span> }
+        </React.Fragment>
+    )
 }
